+++ scripts/run_offline_regression.sh
@@ -380,6 +380,7 @@ swiftc \
   "$CORE/Models/LedgerSyncPlan.swift" \
   "$CORE/Models/LedgerDashboardCloudSnapshot.swift" \
   "$CORE/Models/HotelStay.swift" \
+  "$CORE/Models/PendingAction.swift" \
   "$CORE/Models/ProAccessPolicy.swift" \
   "$CORE/Models/Transaction.swift" \
   "$CORE/Models/Subscription.swift" \